server/requirements.txt
paddleocr==2.8.1
flask
flask-cors
opencv-python-headless
numpy
Pillow
paddlepaddle